+++ src/aosm/azext_aosm/_client_factory.py
@@ -5,6 +5,7 @@
 
 from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.profiles import ResourceType
+
 from .vendored_sdks import HybridNetworkManagementClient