+++ src/aosm/azext_aosm/custom.py
@@ -8,29 +8,31 @@ import os
 import shutil
 from dataclasses import asdict
 from typing import Optional
-from knack.log import get_logger
+
 from azure.cli.core.azclierror import (
     CLIInternalError,
     InvalidArgumentValueError,
     UnclassifiedUserFault,
 )
+from knack.log import get_logger
 
-from azext_aosm.generate_nfd.cnf_nfd_generator import CnfNfdGenerator
-from azext_aosm.generate_nfd.nfd_generator_base import NFDGenerator
-from azext_aosm.generate_nsd.nsd_generator import NSDGenerator
-from azext_aosm.generate_nfd.vnf_nfd_generator import VnfNfdGenerator
-from azext_aosm.delete.delete import ResourceDeleter
-from azext_aosm.deploy.deploy_with_arm import DeployerViaArm
-from azext_aosm.util.constants import VNF, CNF, NSD
-from azext_aosm.util.management_clients import ApiClients
-from azext_aosm.vendored_sdks import HybridNetworkManagementClient
 from azext_aosm._client_factory import cf_resources
 from azext_aosm._configuration import (
-    get_configuration,
+    CNFConfiguration,
     NFConfiguration,
     NSConfiguration,
+    VNFConfiguration,
+    get_configuration,
 )
-
+from azext_aosm.delete.delete import ResourceDeleter
+from azext_aosm.deploy.deploy_with_arm import DeployerViaArm
+from azext_aosm.generate_nfd.cnf_nfd_generator import CnfNfdGenerator
+from azext_aosm.generate_nfd.nfd_generator_base import NFDGenerator
+from azext_aosm.generate_nfd.vnf_nfd_generator import VnfNfdGenerator
+from azext_aosm.generate_nsd.nsd_generator import NSDGenerator
+from azext_aosm.util.constants import CNF, NSD, VNF
+from azext_aosm.util.management_clients import ApiClients
+from azext_aosm.vendored_sdks import HybridNetworkManagementClient
 
 logger = get_logger(__name__)
 
@@ -38,6 +40,8 @@ logger = get_logger(__name__)
 def build_definition(
     definition_type: str,
     config_file: str,
+    order_params: bool = False,
+    interactive: bool = False,
 ):
     """
     Build a definition.
@@ -54,7 +58,12 @@ def build_definition(
     )
 
     # Generate the NFD and the artifact manifest.
-    _generate_nfd(definition_type=definition_type, config=config)
+    _generate_nfd(
+        definition_type=definition_type,
+        config=config,
+        order_params=order_params,
+        interactive=interactive,
+    )
 
 
 def generate_definition_config(definition_type: str, output_file: str = "input.json"):
@@ -90,13 +99,17 @@ def _get_config_from_file(
     return config
 
 
-def _generate_nfd(definition_type, config):
+def _generate_nfd(
+    definition_type: str, config: NFConfiguration, order_params: bool, interactive: bool
+):
     """Generate a Network Function Definition for the given type and config."""
     nfd_generator: NFDGenerator
     if definition_type == VNF:
-        nfd_generator = VnfNfdGenerator(config)
+        assert isinstance(config, VNFConfiguration)
+        nfd_generator = VnfNfdGenerator(config, order_params, interactive)
     elif definition_type == CNF:
-        nfd_generator = CnfNfdGenerator(config)
+        assert isinstance(config, CNFConfiguration)
+        nfd_generator = CnfNfdGenerator(config, interactive)
     else:
         raise CLIInternalError(
             "Generate NFD called for unrecognised definition_type. Only VNF and CNF have been implemented."
@@ -199,6 +212,7 @@ def delete_published_definition(
 def generate_design_config(output_file: str = "input.json"):
     """
     Generate an example config file for building a NSD.
+
     :param output_file: path to output config file, defaults to "input.json"
     :type output_file: str, optional
     """
@@ -208,6 +222,7 @@ def generate_design_config(output_file: str = "input.json"):
 def _generate_config(configuration_type: str, output_file: str = "input.json"):
     """
     Generic generate config function for NFDs and NSDs.
+
     :param configuration_type: CNF, VNF or NSD
     :param output_file: path to output config file, defaults to "input.json"
     :type output_file: str, optional
@@ -224,7 +239,7 @@ def _generate_config(configuration_type: str, output_file: str = "input.json"):
 
     with open(output_file, "w", encoding="utf-8") as f:
         f.write(config_as_dict)
-        if configuration_type == CNF or configuration_type == VNF:
+        if configuration_type in (CNF,VNF):
             prtName = "definition"
         else:
             prtName = "design"
@@ -235,6 +250,7 @@ def _generate_config(configuration_type: str, output_file: str = "input.json"):
 def build_design(cmd, client: HybridNetworkManagementClient, config_file: str):
     """
     Build a Network Service Design.
+
     :param cmd:
     :type cmd: _type_
     :param client:
@@ -266,6 +282,7 @@ def delete_published_design(
 ):
     """
     Delete a published NSD.
+
     :param config_file: Path to the config file
     :param clean: if True, will delete the NSDG, artifact stores and publisher too.
                   Defaults to False. Only works if no resources have those as a parent.
@@ -292,6 +309,7 @@ def publish_design(
 ):
     """
     Publish a generated design.
+
     :param cmd:
     :param client:
     :type client: HybridNetworkManagementClient
@@ -331,8 +349,6 @@ def _generate_nsd(config: NSDGenerator, api_clients):
     if config:
         nsd_generator = NSDGenerator(config)
     else:
-        from azure.cli.core.azclierror import CLIInternalError
-
         raise CLIInternalError("Generate NSD called without a config file")
     deploy_parameters = _get_nfdv_deployment_parameters(config, api_clients)