+++ src/aosm/azext_aosm/deploy/deploy_with_arm.py
@@ -7,26 +7,26 @@ import json
 import os
 import shutil
 import subprocess  # noqa
-from typing import Any, Dict, Optional
 import tempfile
 import time
+from typing import Any, Dict, Optional
 
-from knack.log import get_logger
 from azure.mgmt.resource.resources.models import DeploymentExtended
+from knack.log import get_logger
 
-from azext_aosm.deploy.artifact_manifest import ArtifactManifestOperator
-from azext_aosm.util.management_clients import ApiClients
-from azext_aosm.deploy.pre_deploy import PreDeployerViaSDK
 from azext_aosm._configuration import NFConfiguration, NSConfiguration, VNFConfiguration
+from azext_aosm.deploy.artifact_manifest import ArtifactManifestOperator
+from azext_aosm.deploy.pre_deploy import PreDeployerViaSDK
 from azext_aosm.util.constants import (
-    NSD_DEFINITION_BICEP_FILE,
-    NSD_ARTIFACT_MANIFEST_BICEP_FILE,
     NF_DEFINITION_BICEP_FILE,
+    NSD,
+    NSD_ARTIFACT_MANIFEST_BICEP_FILE,
+    NSD_DEFINITION_BICEP_FILE,
+    VNF,
     VNF_DEFINITION_BICEP_TEMPLATE,
     VNF_MANIFEST_BICEP_TEMPLATE,
-    NSD,
-    VNF,
 )
+from azext_aosm.util.management_clients import ApiClients
 
 logger = get_logger(__name__)
 
@@ -193,7 +193,7 @@ class DeployerViaArm:
                 "saArtifactStoreName": {"value": self.config.blob_artifact_store_name},
                 "acrManifestName": {"value": self.config.acr_manifest_name},
                 "saManifestName": {"value": self.config.sa_manifest_name},
-                'nfName': {"value": self.config.nf_name},
+                "nfName": {"value": self.config.nf_name},
                 "vhdVersion": {"value": self.config.vhd.version},
                 "armTemplateVersion": {"value": self.config.arm_template.version},
             }
@@ -427,11 +427,15 @@ class DeployerViaArm:
             # Validation failed so don't even try to deploy
             logger.error(
                 "Template for resource group %s has failed validation. The message was: %s.\
-                See logs for additional details.", resource_group, validation_res.error.message
+                See logs for additional details.",
+                resource_group,
+                validation_res.error.message,
             )
             logger.debug(
                 "Template for resource group %s failed validation. \
-                Full error details: %s", resource_group, validation_res.error
+                Full error details: %s",
+                resource_group,
+                validation_res.error,
             )
             raise RuntimeError("Azure template validation failed.")
 
@@ -470,7 +474,9 @@ class DeployerViaArm:
                 f"\nAborting"
             )
         logger.debug(
-            "Provisioning state of deployment %s : %s", resource_group, depl_props.provisioning_state
+            "Provisioning state of deployment %s : %s",
+            resource_group,
+            depl_props.provisioning_state,
         )
 
         return depl_props.outputs