+++ src/aosm/azext_aosm/generate_nfd/cnf_nfd_generator.py
@@ -8,29 +8,33 @@ import os
 import re
 import shutil
 import tarfile
-from typing import Dict, List, Any, Tuple, Optional, Iterator
-
 import tempfile
+from typing import Any, Dict, Iterator, List, Optional, Tuple
+
 import yaml
-from jinja2 import Template, StrictUndefined
-from azure.cli.core.azclierror import InvalidTemplateError, FileOperationError
+from azure.cli.core.azclierror import FileOperationError, InvalidTemplateError
+from jinja2 import StrictUndefined, Template
 from knack.log import get_logger
 
-from azext_aosm.generate_nfd.nfd_generator_base import NFDGenerator
 from azext_aosm._configuration import CNFConfiguration, HelmPackageConfig
+from azext_aosm.generate_nfd.nfd_generator_base import NFDGenerator
 from azext_aosm.util.constants import (
     CNF_DEFINITION_BICEP_TEMPLATE,
     CNF_DEFINITION_JINJA2_SOURCE_TEMPLATE,
     CNF_MANIFEST_BICEP_TEMPLATE,
     CNF_MANIFEST_JINJA2_SOURCE_TEMPLATE,
+    CONFIG_MAPPINGS,
     DEPLOYMENT_PARAMETER_MAPPING_REGEX,
     IMAGE_NAME_AND_VERSION_REGEX,
     IMAGE_PATH_REGEX,
     CONFIG_MAPPINGS,
     SCHEMAS,
-    SCHEMA_PREFIX,
     DEPLOYMENT_PARAMETERS,
+    GENERATED_VALUES_MAPPINGS,
+    SCHEMA_PREFIX,
+    SCHEMAS,
 )
+from azext_aosm.util.utils import input_ack
 
 logger = get_logger(__name__)
 
@@ -46,8 +50,14 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
     - A bicep file for the Artifact manifests
     """
 
-    def __init__(self, config: CNFConfiguration):
-        """Create a new CNF NFD Generator."""
+    def __init__(self, config: CNFConfiguration, interactive: bool = False):
+        """
+        Create a new CNF NFD Generator.
+
+        Interactive parameter is only used if the user wants to generate the values
+        mapping file from the values.yaml in the helm package, and also requires the
+        mapping file in config to be blank.
+        """
         super(NFDGenerator, self).__init__()
         self.config = config
         self.nfd_jinja2_template_path = os.path.join(
@@ -69,7 +79,8 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         self._bicep_path = os.path.join(
             self.output_folder_name, CNF_DEFINITION_BICEP_TEMPLATE
         )
-        self._tmp_folder_name = ''
+        self.interactive = interactive
+        self._tmp_folder_name = ""
 
     def generate_nfd(self) -> None:
         """Generate a CNF NFD which comprises a group, an Artifact Manifest and an NFDV."""
@@ -86,6 +97,10 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
                     self._extract_chart(helm_package.path_to_chart)
 
                     # TODO: Validate charts
+
+                    # Create a chart mapping schema if none has been passed in.
+                    if not helm_package.path_to_mappings:
+                        self._generate_chart_value_mappings(helm_package)
 
                     # Get schema for each chart
                     # (extract mappings and take the schema bits we need from values.schema.json)
@@ -157,7 +172,7 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         logger.debug("Extracting helm package %s", path)
 
         (_, ext) = os.path.splitext(path)
-        if ext in ('.gz', '.tgz'):
+        if ext in (".gz", ".tgz"):
             with tarfile.open(path, "r:gz") as tar:
                 tar.extractall(path=self._tmp_folder_name)
 
@@ -170,6 +185,64 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
                 f"ERROR: The helm package '{path}' is not a .tgz, .tar or .tar.gz file.\
                 Please fix this and run the command again."
             )
+
+    def _generate_chart_value_mappings(self, helm_package: HelmPackageConfig) -> None:
+        """
+        Optional function to create a chart value mappings file with every value being a deployParameter.
+
+        Expected use when a helm chart is very simple and user wants every value to be a
+        deployment parameter.
+        """
+        logger.debug(
+            "Creating chart value mappings file for %s", helm_package.path_to_chart
+        )
+        print("Creating chart value mappings file for %s", helm_package.path_to_chart)
+
+        # Get all the values files in the chart
+        top_level_values_yaml = self._read_top_level_values_yaml(helm_package)
+
+        mapping_to_write = self._replace_values_with_deploy_params(
+            top_level_values_yaml, {}
+        )
+
+        # Write the mapping to a file
+        folder_name = os.path.join(self._tmp_folder_name, GENERATED_VALUES_MAPPINGS)
+        os.makedirs(folder_name, exist_ok=True)
+        mapping_filepath = os.path.join(
+            self._tmp_folder_name,
+            GENERATED_VALUES_MAPPINGS,
+            f"{helm_package.name}-generated-mapping.yaml",
+        )
+        with open(mapping_filepath, "w", encoding="UTF-8") as mapping_file:
+            yaml.dump(mapping_to_write, mapping_file)
+
+        # Update the config that points to the mapping file
+        helm_package.path_to_mappings = mapping_filepath
+
+    def _read_top_level_values_yaml(
+        self, helm_package: HelmPackageConfig
+    ) -> Dict[str, Any]:
+        """Return a dictionary of the values.yaml|yml read from the root of the helm package.
+
+        :param helm_package: The helm package to look in
+        :type helm_package: HelmPackageConfig
+        :raises FileOperationError: if no values.yaml|yml found
+        :return: A dictionary of the yaml read from the file
+        :rtype: Dict[str, Any]
+        """
+        for file in os.listdir(os.path.join(self._tmp_folder_name, helm_package.name)):
+            if file in ("values.yaml", "values.yml"):
+                with open(
+                    os.path.join(self._tmp_folder_name, helm_package.name, file),
+                    "r",
+                    encoding="UTF-8",
+                ) as values_file:
+                    values_yaml = yaml.safe_load(values_file)
+                return values_yaml
+
+        raise FileOperationError(
+            "Cannot find top level values.yaml/.yml file in Helm package."
+        )
 
     def write_manifest_bicep_file(self) -> None:
         """Write the bicep file for the Artifact Manifest."""
@@ -227,6 +300,7 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         os.mkdir(self.output_folder_name)
         os.mkdir(os.path.join(self.output_folder_name, SCHEMAS))
 
+        # Copy the nfd and the manifest bicep files to the output folder
         tmp_nfd_bicep_path = os.path.join(
             self._tmp_folder_name, CNF_DEFINITION_BICEP_TEMPLATE
         )
@@ -236,7 +310,23 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
             self._tmp_folder_name, CNF_MANIFEST_BICEP_TEMPLATE
         )
         shutil.copy(tmp_manifest_bicep_path, self.output_folder_name)
+        
+        # Copy any generated values mappings YAML files to the corresponding folder in
+        # the output directory so that the user can edit them and re-run the build if
+        # required
+        if os.path.exists(
+            os.path.join(self._tmp_folder_name, GENERATED_VALUES_MAPPINGS)
+        ):
+            generated_mappings_path = os.path.join(
+                self.output_folder_name, GENERATED_VALUES_MAPPINGS
+            )
+            shutil.copytree(
+                os.path.join(self._tmp_folder_name, GENERATED_VALUES_MAPPINGS),
+                generated_mappings_path,
+            )
 
+        # Copy the JSON config mappings and deploymentParameters schema that are used
+        # for the NFD to the output folder
         tmp_config_mappings_path = os.path.join(self._tmp_folder_name, CONFIG_MAPPINGS)
         output_config_mappings_path = os.path.join(
             self.output_folder_name, CONFIG_MAPPINGS
@@ -277,7 +367,7 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
             "dependsOnProfile": helm_package.depends_on,
             "registryValuesPaths": list(registry_values_paths),
             "imagePullSecretsValuesPaths": list(image_pull_secrets_values_paths),
-            "valueMappingsPath": self.generate_parameter_mappings(helm_package),
+            "valueMappingsPath": self.jsonify_value_mappings(helm_package),
         }
 
     def _find_yaml_files(self, directory) -> Iterator[str]:
@@ -325,8 +415,8 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         """
         Get the list of artifacts for the chart.
 
-        param helm_package: The helm package config.
-        param image_line_matches: The list of image line matches.
+        param helm_package: The helm package config. param image_line_matches: The list
+        of image line matches.
         """
         artifact_list = []
         (chart_name, chart_version) = self.get_chart_name_and_version(helm_package)
@@ -364,7 +454,7 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         if not os.path.exists(mappings_path):
             raise InvalidTemplateError(
                 f"ERROR: The helm package '{helm_package.name}' does not have a valid values mappings file. \
-                    The file at '{helm_package.path_to_mappings}' does not exist. \
+                    The file at '{helm_package.path_to_mappings}' does not exist.\n\
                     Please fix this and run the command again."
             )
         if not os.path.exists(values_schema):
@@ -394,7 +484,32 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
     def find_deploy_params(
         self, nested_dict, schema_nested_dict, final_schema
     ) -> Dict[Any, Any]:
-        """Find the deploy parameters in the values.mappings.yaml file and add them to the schema."""
+        """
+        Create a schema of types of only those values in the values.mappings.yaml file which have a deployParameters mapping.
+        
+        Finds the relevant part of the full schema of the values file and finds the 
+        type of the parameter name, then adds that to the final schema, with no nesting.
+        
+        Returns a schema of form:
+        {
+            "$schema": "https://json-schema.org/draft-07/schema#",
+            "title": "DeployParametersSchema",
+            "type": "object",
+            "properties": {
+                "<parameter1>": {
+                    "type": "<type>"
+                },
+                "<parameter2>": {
+                    "type": "<type>"
+                },
+        
+        nested_dict: the dictionary of the values mappings yaml which contains 
+                     deployParameters mapping placeholders
+        schema_nested_dict: the properties section of the full schema (or sub-object in 
+                            schema)
+        final_schema: Blank dictionary if this is the top level starting point, 
+                      otherwise the final_schema as far as we have got.
+        """
         original_schema_nested_dict = schema_nested_dict
         # if given a blank mapping file, return empty schema
         if nested_dict is None:
@@ -402,28 +517,103 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
         for k, v in nested_dict.items():
             # if value is a string and contains deployParameters.
             if isinstance(v, str) and re.search(DEPLOYMENT_PARAMETER_MAPPING_REGEX, v):
-                # only add the parameter name (e.g. from {deployParameter.zone} only param = zone)
+                logger.debug(
+                    "Found string deploy parameter for key %s, value %s. Find schema type",
+                    k,
+                    v,
+                )
+                # only add the parameter name (e.g. from {deployParameter.zone} only
+                # param = zone)
                 param = v.split(".", 1)[1]
                 param = param.split("}", 1)[0]
 
-                # add the schema for k (from the big schema) to the (smaller) schema
-                final_schema.update(
-                    {param: {"type": schema_nested_dict["properties"][k]["type"]}}
-                )
-
+                # add the schema for k (from the full schema) to the (new) schema
+                if "properties" in schema_nested_dict.keys():
+                    # Occurs if top level item in schema properties is an object with
+                    # properties itself
+                    final_schema.update(
+                        {param: {"type": schema_nested_dict["properties"][k]["type"]}}
+                    )
+                else:
+                    # Occurs if top level schema item in schema properties are objects
+                    # with no "properties" - but can have "type".
+                    final_schema.update(
+                        {param: {"type": schema_nested_dict[k]["type"]}}
+                    )
             # else if value is a (non-empty) dictionary (i.e another layer of nesting)
             elif hasattr(v, "items") and v.items():
-                # handling schema having properties which doesn't map directly to the values file nesting
+                logger.debug("Found dict value for key %s. Find schema type", k)
+                # handling schema having properties which doesn't map directly to the
+                # values file nesting
                 if "properties" in schema_nested_dict.keys():
                     schema_nested_dict = schema_nested_dict["properties"][k]
                 else:
                     schema_nested_dict = schema_nested_dict[k]
                 # recursively call function with values (i.e the nested dictionary)
                 self.find_deploy_params(v, schema_nested_dict, final_schema)
-                # reset the schema dict to its original value (once finished with that level of recursion)
+                # reset the schema dict to its original value (once finished with that
+                # level of recursion)
                 schema_nested_dict = original_schema_nested_dict
 
         return final_schema
+
+    def _replace_values_with_deploy_params(
+        self,
+        values_yaml_dict,
+        param_prefix: Optional[str] = None,
+    ) -> Dict[Any, Any]:
+        """
+        Given the yaml dictionary read from values.yaml, replace all the values with {deploymentParameter.keyname}.
+
+        Thus creating a values mapping file if the user has not provided one in config.
+        """
+        logger.debug("Replacing values with deploy parameters")
+        final_values_mapping_dict: Dict[Any, Any] = {}
+        for k, v in values_yaml_dict.items():
+            # if value is a string and contains deployParameters.
+            logger.debug("Processing key %s", k)
+            param_name = k if param_prefix is None else f"{param_prefix}_{k}"
+            if isinstance(v, (str, int, bool)):
+                # Replace the parameter with {deploymentParameter.keyname}
+                if self.interactive:
+                    # Interactive mode. Prompt user to include or exclude parameters
+                    # This requires the enter key after the y/n input which isn't ideal
+                    if not input_ack("y", f"Expose parameter {param_name}? y/n "):
+                        logger.debug("Excluding parameter %s", param_name)
+                        final_values_mapping_dict.update({k: v})
+                        continue
+                replacement_value = f"{{deployParameters.{param_name}}}"
+
+                # add the schema for k (from the big schema) to the (smaller) schema
+                final_values_mapping_dict.update({k: replacement_value})
+            elif isinstance(v, dict):
+                
+                final_values_mapping_dict[k] = self._replace_values_with_deploy_params(
+                    v, param_name
+                )
+            elif isinstance(v, list):
+                final_values_mapping_dict[k] = []
+                for index, item in enumerate(v):
+                    param_name = f"{param_prefix}_{k}_{index}" if param_prefix else f"{k})_{index}"        
+                    if isinstance(item, dict):
+                        final_values_mapping_dict[k].append(
+                            self._replace_values_with_deploy_params(
+                                item, param_name
+                            )
+                        )
+                    elif isinstance(v, (str, int, bool)):
+                        replacement_value = f"{{deployParameters.{param_name}}}"
+                        final_values_mapping_dict[k].append(replacement_value)
+                    else:
+                        raise ValueError(
+                            f"Found an unexpected type {type(v)} of key {k} in "
+                            "values.yaml, cannot generate values mapping file.")                        
+            else:
+                raise ValueError(
+                    f"Found an unexpected type {type(v)} of key {k} in values.yaml, "
+                    "cannot generate values mapping file.")
+
+        return final_values_mapping_dict
 
     def get_chart_name_and_version(
         self, helm_package: HelmPackageConfig
@@ -450,12 +640,10 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
 
         return (chart_name, chart_version)
 
-    def generate_parameter_mappings(self, helm_package: HelmPackageConfig) -> str:
-        """Generate parameter mappings for the given helm package."""
-        # values = os.path.join(
-        #     self._tmp_folder_name, helm_package.name, "values.mappings.yaml"
-        # )
-        values = helm_package.path_to_mappings
+    def jsonify_value_mappings(self, helm_package: HelmPackageConfig) -> str:
+        """Yaml->JSON values mapping file, then return path to it."""
+        mappings_yaml = helm_package.path_to_mappings
+
         mappings_folder_path = os.path.join(self._tmp_folder_name, CONFIG_MAPPINGS)
         mappings_filename = f"{helm_package.name}-mappings.json"
 
@@ -464,7 +652,7 @@ class CnfNfdGenerator(NFDGenerator):  # pylint: disable=too-many-instance-attrib
 
         mapping_file_path = os.path.join(mappings_folder_path, mappings_filename)
 
-        with open(values, "r", encoding="utf-8") as f:
+        with open(mappings_yaml, "r", encoding="utf-8") as f:
             data = yaml.load(f, Loader=yaml.FullLoader)
 
         with open(mapping_file_path, "w", encoding="utf-8") as file: