+++ src/aosm/azext_aosm/util/management_clients.py
@@ -4,8 +4,9 @@
 # --------------------------------------------------------------------------------------------
 """Clients for the python SDK along with useful caches."""
 
-from knack.log import get_logger
 from azure.mgmt.resource import ResourceManagementClient
+from knack.log import get_logger
+
 from azext_aosm.vendored_sdks import HybridNetworkManagementClient
 
 logger = get_logger(__name__)